+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/in_app_webview/InAppWebView.java
@@ -9,6 +9,7 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Point;
 import android.graphics.drawable.ColorDrawable;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
@@ -22,7 +23,6 @@ import android.util.AttributeSet;
 import android.util.Log;
 import android.view.ActionMode;
 import android.view.ContextMenu;
-import android.view.DragEvent;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -40,6 +40,7 @@ import android.webkit.DownloadListener;
 import android.webkit.ValueCallback;
 import android.webkit.WebBackForwardList;
 import android.webkit.WebHistoryItem;
+import android.webkit.WebMessage;
 import android.webkit.WebSettings;
 import android.webkit.WebStorage;
 import android.widget.HorizontalScrollView;
@@ -53,6 +54,7 @@ import androidx.webkit.WebViewCompat;
 import androidx.webkit.WebViewFeature;
 
 import com.pichillilorenzo.flutter_inappwebview.InAppWebViewFlutterPlugin;
+import com.pichillilorenzo.flutter_inappwebview.types.InAppWebViewInterface;
 import com.pichillilorenzo.flutter_inappwebview.JavaScriptBridgeInterface;
 import com.pichillilorenzo.flutter_inappwebview.R;
 import com.pichillilorenzo.flutter_inappwebview.Util;
@@ -99,7 +101,7 @@ import okhttp3.OkHttpClient;
 import static android.content.Context.INPUT_METHOD_SERVICE;
 import static com.pichillilorenzo.flutter_inappwebview.types.PreferredContentModeOptionType.fromValue;
 
-final public class InAppWebView extends InputAwareWebView {
+final public class InAppWebView extends InputAwareWebView implements InAppWebViewInterface {
 
   static final String LOG_TAG = "InAppWebView";
 
@@ -129,7 +131,7 @@ final public class InAppWebView extends InputAwareWebView {
   public LinearLayout floatingContextMenu = null;
   @Nullable
   public Map<String, Object> contextMenu = null;
-  public Handler headlessHandler = new Handler(Looper.getMainLooper());
+  public Handler mainLooperHandler = new Handler(Looper.getMainLooper());
   static Handler mHandler = new Handler();
 
   public Runnable checkScrollStoppedTask;
@@ -145,6 +147,7 @@ final public class InAppWebView extends InputAwareWebView {
   public Map<String, ValueCallback<String>> evaluateJavaScriptContentWorldCallbacks = new HashMap<>();
 
   public Map<String, WebMessageChannel> webMessageChannels = new HashMap<>();
+  public List<WebMessageListener> webMessageListeners = new ArrayList<>();
 
   public InAppWebView(Context context) {
     super(context);
@@ -172,11 +175,6 @@ final public class InAppWebView extends InputAwareWebView {
     this.contextMenu = contextMenu;
     this.userContentController.addUserOnlyScripts(userScripts);
     plugin.activity.registerForContextMenu(this);
-  }
-
-  @Override
-  public void reload() {
-    super.reload();
   }
 
   public void prepare() {
@@ -400,7 +398,7 @@ final public class InAppWebView extends InputAwareWebView {
           onScrollStopped();
         } else {
           initialPositionScrollStoppedTask = getScrollY();
-          headlessHandler.postDelayed(checkScrollStoppedTask, newCheckScrollStoppedTask);
+          mainLooperHandler.postDelayed(checkScrollStoppedTask, newCheckScrollStoppedTask);
         }
       }
     };
@@ -418,7 +416,7 @@ final public class InAppWebView extends InputAwareWebView {
                     hideContextMenu();
                   }
                 } else {
-                  headlessHandler.postDelayed(checkContextMenuShouldBeClosedTask, newCheckContextMenuShouldBeClosedTaskTask);
+                  mainLooperHandler.postDelayed(checkContextMenuShouldBeClosedTask, newCheckContextMenuShouldBeClosedTaskTask);
                 }
               }
             });
@@ -471,11 +469,9 @@ final public class InAppWebView extends InputAwareWebView {
     setOnLongClickListener(new OnLongClickListener() {
       @Override
       public boolean onLongClick(View v) {
-        HitTestResult hitTestResult = getHitTestResult();
-        Map<String, Object> obj = new HashMap<>();
-        obj.put("type", hitTestResult.getType());
-        obj.put("extra", hitTestResult.getExtra());
-        channel.invokeMethod("onLongPressHitTestResult", obj);
+        com.pichillilorenzo.flutter_inappwebview.types.HitTestResult hitTestResult =
+                com.pichillilorenzo.flutter_inappwebview.types.HitTestResult.fromWebViewHitTestResult(getHitTestResult());
+        channel.invokeMethod("onLongPressHitTestResult", hitTestResult.toMap());
         return false;
       }
     });
@@ -540,6 +536,10 @@ final public class InAppWebView extends InputAwareWebView {
   }
 
   public void loadFile(String assetFilePath) throws IOException {
+    if (plugin == null) {
+      return;
+    }
+    
     loadUrl(Util.getUrlAsset(plugin, assetFilePath));
   }
 
@@ -570,12 +570,12 @@ final public class InAppWebView extends InputAwareWebView {
   public void takeScreenshot(final @Nullable Map<String, Object> screenshotConfiguration, final MethodChannel.Result result) {
     final float pixelDensity = Util.getPixelDensity(getContext());
     
-    headlessHandler.post(new Runnable() {
+    mainLooperHandler.post(new Runnable() {
       @Override
       public void run() {
         try {
-          Bitmap resized = Bitmap.createBitmap(getMeasuredWidth(), getMeasuredHeight(), Bitmap.Config.ARGB_8888);
-          Canvas c = new Canvas(resized);
+          Bitmap screenshotBitmap = Bitmap.createBitmap(getMeasuredWidth(), getMeasuredHeight(), Bitmap.Config.ARGB_8888);
+          Canvas c = new Canvas(screenshotBitmap);
           c.translate(-getScrollX(), -getScrollY());
           draw(c);
 
@@ -588,10 +588,10 @@ final public class InAppWebView extends InputAwareWebView {
             if (rect != null) {
               int rectX = (int) Math.floor(rect.get("x") * pixelDensity + 0.5);
               int rectY = (int) Math.floor(rect.get("y") * pixelDensity + 0.5);
-              int rectWidth = Math.min(resized.getWidth(), (int) Math.floor(rect.get("width") * pixelDensity + 0.5));
-              int rectHeight = Math.min(resized.getHeight(), (int) Math.floor(rect.get("height") * pixelDensity + 0.5));
-              resized = Bitmap.createBitmap(
-                      resized,
+              int rectWidth = Math.min(screenshotBitmap.getWidth(), (int) Math.floor(rect.get("width") * pixelDensity + 0.5));
+              int rectHeight = Math.min(screenshotBitmap.getHeight(), (int) Math.floor(rect.get("height") * pixelDensity + 0.5));
+              screenshotBitmap = Bitmap.createBitmap(
+                      screenshotBitmap,
                       rectX,
                       rectY,
                       rectWidth,
@@ -601,9 +601,9 @@ final public class InAppWebView extends InputAwareWebView {
             Double snapshotWidth = (Double) screenshotConfiguration.get("snapshotWidth");
             if (snapshotWidth != null) {
               int dstWidth = (int) Math.floor(snapshotWidth * pixelDensity + 0.5);
-              float ratioBitmap = (float) resized.getWidth() / (float) resized.getHeight();
+              float ratioBitmap = (float) screenshotBitmap.getWidth() / (float) screenshotBitmap.getHeight();
               int dstHeight = (int) ((float) dstWidth / ratioBitmap);
-              resized = Bitmap.createScaledBitmap(resized, dstWidth, dstHeight, true);
+              screenshotBitmap = Bitmap.createScaledBitmap(screenshotBitmap, dstWidth, dstHeight, true);
             }
 
             try {
@@ -615,7 +615,7 @@ final public class InAppWebView extends InputAwareWebView {
             quality = (Integer) screenshotConfiguration.get("quality");
           }
 
-          resized.compress(
+          screenshotBitmap.compress(
                   compressFormat,
                   quality,
                   byteArrayOutputStream);
@@ -625,7 +625,7 @@ final public class InAppWebView extends InputAwareWebView {
           } catch (IOException e) {
             e.printStackTrace();
           }
-          resized.recycle();
+          screenshotBitmap.recycle();
           result.success(byteArrayOutputStream.toByteArray());
 
         } catch (IllegalArgumentException e) {
@@ -982,7 +982,7 @@ final public class InAppWebView extends InputAwareWebView {
               .replace(PluginScriptsUtil.VAR_RESULT_UUID, resultUuid);
     }
     final String finalScriptToInject = scriptToInject;
-    headlessHandler.post(new Runnable() {
+    mainLooperHandler.post(new Runnable() {
       @Override
       public void run() {
         String scriptToInject = userContentController.generateCodeForScriptEvaluation(finalScriptToInject, contentWorld);
@@ -1231,11 +1231,9 @@ final public class InAppWebView extends InputAwareWebView {
   }
 
   private void sendOnCreateContextMenuEvent() {
-    HitTestResult hitTestResult = getHitTestResult();
-    Map<String, Object> obj = new HashMap<>();
-    obj.put("type", hitTestResult.getType());
-    obj.put("extra", hitTestResult.getExtra());
-    channel.invokeMethod("onCreateContextMenu", obj);
+    com.pichillilorenzo.flutter_inappwebview.types.HitTestResult hitTestResult =
+            com.pichillilorenzo.flutter_inappwebview.types.HitTestResult.fromWebViewHitTestResult(getHitTestResult());
+    channel.invokeMethod("onCreateContextMenu", hitTestResult.toMap());
   }
 
   private Point contextMenuPoint = new Point(0, 0);
@@ -1477,7 +1475,7 @@ final public class InAppWebView extends InputAwareWebView {
             new LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT, curx, ((int) cury) + getScrollY())
     );
 
-    headlessHandler.post(new Runnable() {
+    mainLooperHandler.post(new Runnable() {
       @Override
       public void run() {
         if (floatingContextMenu != null) {
@@ -1547,16 +1545,6 @@ final public class InAppWebView extends InputAwareWebView {
       public void onReceiveValue(String value) {
         value = (value != null && !value.equalsIgnoreCase("null")) ? value.substring(1, value.length() - 1) : null;
         resultCallback.onReceiveValue(value);
-      }
-    });
-  }
-
-  @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-  public void getSelectedText(final MethodChannel.Result result) {
-    getSelectedText(new ValueCallback<String>() {
-      @Override
-      public void onReceiveValue(String value) {
-        result.success(value);
       }
     });
   }
@@ -1649,8 +1637,16 @@ final public class InAppWebView extends InputAwareWebView {
     return webMessageChannel;
   }
 
-  public void addWebMessageListener(@NonNull WebMessageListener webMessageListener) {
+  @Override
+  public WebMessageChannel createWebMessageChannel(ValueCallback<WebMessageChannel> callback) {
+    WebMessageChannel webMessageChannel = createCompatWebMessageChannel();
+    callback.onReceiveValue(webMessageChannel);
+    return webMessageChannel;
+  }
+
+  public void addWebMessageListener(@NonNull WebMessageListener webMessageListener) throws Exception {
     WebViewCompat.addWebMessageListener(this, webMessageListener.jsObjectName, webMessageListener.allowedOriginRules, webMessageListener.listener);
+    webMessageListeners.add(webMessageListener);
   }
 
   public void disposeWebMessageChannels() {
@@ -1660,19 +1656,94 @@ final public class InAppWebView extends InputAwareWebView {
     webMessageChannels.clear();
   }
 
-//  @Override
+  public void disposeWebMessageListeners() {
+    for (WebMessageListener webMessageListener : webMessageListeners) {
+      webMessageListener.dispose();
+    }
+    webMessageListeners.clear();
+  }
+
+  @Override
+  public void postWebMessage(com.pichillilorenzo.flutter_inappwebview.types.WebMessage message, Uri targetOrigin, ValueCallback<String> callback) throws Exception {
+    throw new UnsupportedOperationException();
+  }
+
+  //  @Override
 //  protected void onWindowVisibilityChanged(int visibility) {
 //    if (visibility != View.GONE) super.onWindowVisibilityChanged(View.VISIBLE);
 //  }
+
+  public float getZoomScale() {
+    return zoomScale;
+  }
+
+  @Override
+  public void getZoomScale(final ValueCallback<Float> callback) {
+    callback.onReceiveValue(zoomScale);
+  }
+
+  @Nullable
+  public Map<String, Object> getContextMenu() {
+    return contextMenu;
+  }
+
+  public void setContextMenu(@Nullable Map<String, Object> contextMenu) {
+    this.contextMenu = contextMenu;
+  }
+
+  @Nullable
+  public InAppWebViewFlutterPlugin getPlugin() {
+    return plugin;
+  }
+
+  public void setPlugin(@Nullable InAppWebViewFlutterPlugin plugin) {
+    this.plugin = plugin;
+  }
+
+  @Nullable
+  public InAppBrowserDelegate getInAppBrowserDelegate() {
+    return inAppBrowserDelegate;
+  }
+
+  public void setInAppBrowserDelegate(@Nullable InAppBrowserDelegate inAppBrowserDelegate) {
+    this.inAppBrowserDelegate = inAppBrowserDelegate;
+  }
+
+  public UserContentController getUserContentController() {
+    return userContentController;
+  }
+
+  public void setUserContentController(UserContentController userContentController) {
+    this.userContentController = userContentController;
+  }
+
+  public Map<String, WebMessageChannel> getWebMessageChannels() {
+    return webMessageChannels;
+  }
+
+  public void setWebMessageChannels(Map<String, WebMessageChannel> webMessageChannels) {
+    this.webMessageChannels = webMessageChannels;
+  }
+
+  @Override
+  public void getContentHeight(ValueCallback<Integer> callback) {
+    callback.onReceiveValue(getContentHeight());
+  }
+
+  @Override
+  public void getHitTestResult(ValueCallback<com.pichillilorenzo.flutter_inappwebview.types.HitTestResult> callback) {
+    callback.onReceiveValue(com.pichillilorenzo.flutter_inappwebview.types.HitTestResult.fromWebViewHitTestResult(getHitTestResult()));
+  }
 
   @Override
   public void dispose() {
     if (windowId != null) {
       InAppWebViewChromeClient.windowWebViewMessages.remove(windowId);
     }
-    headlessHandler.removeCallbacksAndMessages(null);
+    mainLooperHandler.removeCallbacksAndMessages(null);
     mHandler.removeCallbacksAndMessages(null);
     disposeWebMessageChannels();
+    disposeWebMessageListeners();
     removeAllViews();
     if (checkContextMenuShouldBeClosedTask != null)
       removeCallbacks(checkContextMenuShouldBeClosedTask);