+++ ios/Classes/InAppWebView/InAppWebView.swift
@@ -19,6 +19,7 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
     var pullToRefreshControl: PullToRefreshControl?
     var webMessageChannels: [String:WebMessageChannel] = [:]
     var webMessageListeners: [WebMessageListener] = []
+    var currentOriginalUrl: URL?
     
     static var sslCertificatesMap: [String: SslCertificate] = [:] // [URL host name : SslCertificate]
     static var credentialsProposed: [URLCredential] = []
@@ -26,6 +27,7 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
     var lastScrollX: CGFloat = 0
     var lastScrollY: CGFloat = 0
     
+    // Used to manage pauseTimers() and resumeTimers()
     var isPausedTimers = false
     var isPausedTimersCompletionHandler: (() -> Void)?
 
@@ -1557,6 +1559,9 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
     }
     
     public func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
+        currentOriginalUrl = url
+        lastTouchPoint = nil
+        
         disposeWebMessageChannels()
         initializeWindowIdJS()
         
@@ -2723,6 +2728,10 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
         scrollView.setZoomScale(currentZoomScale * CGFloat(zoomFactor), animated: animated)
     }
     
+    public func getOriginalUrl() -> URL? {
+        return currentOriginalUrl
+    }
+    
     public func getZoomScale() -> Float {
         return Float(scrollView.zoomScale)
     }
@@ -2883,10 +2892,7 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
     }
     
     public func dispose() {
-        if isPausedTimers, let completionHandler = isPausedTimersCompletionHandler {
-            isPausedTimersCompletionHandler = nil
-            completionHandler()
-        }
+        resumeTimers()
         stopLoading()
         disposeWebMessageChannels()
         for webMessageListener in webMessageListeners {